extract components and startinitial documentation
