+++ Cargo.toml
@@ -45,7 +45,7 @@ hyper = { version = "0.14", features = ["full"] }
 hyper-tls = "0.5.0"
 itertools = "0.10.3"
 lazy_static = "1.4.0"
-libp2p = { version = "0.43.0", features=["tcp-tokio"]}
+libp2p = { version = "0.43.0", features=["dns-tokio", "tcp-tokio"]}
 log = { version = "0.4.14", features = ["max_level_trace", "release_max_level_trace"] }
 maplit = "1.0.2"
 num-traits = "0.2.14"
