+++ pyrsia_node/src/network/handlers.rs
@@ -14,22 +14,23 @@
    limitations under the License.
 */
 
-use crate::artifacts_repository::hash_util::HashAlgorithm;
-use crate::network::p2p;
-use crate::node_manager::handlers::{get_artifact, get_artifact_hashes};
 use libp2p::request_response::ResponseChannel;
 use libp2p::Multiaddr;
 use log::{debug, info};
+use pyrsia::artifacts_repository::hash_util::HashAlgorithm;
+use pyrsia::network::artifact_protocol::ArtifactResponse;
+use pyrsia::network::client::Client;
+use pyrsia::node_manager;
 
 /// Reach out to another node with the specified address
-pub async fn dial_other_peer(mut p2p_client: p2p::Client, to_dial: &Multiaddr) {
+pub async fn dial_other_peer(mut p2p_client: Client, to_dial: &Multiaddr) {
     p2p_client.dial(to_dial).await.expect("Dial to succeed.");
     info!("Dialed {:?}", to_dial);
 }
 
 /// Provide all known artifacts on the p2p network
-pub async fn provide_artifacts(mut p2p_client: p2p::Client) {
-    if let Ok(artifact_hashes) = get_artifact_hashes() {
+pub async fn provide_artifacts(mut p2p_client: Client) {
+    if let Ok(artifact_hashes) = node_manager::handlers::get_artifact_hashes() {
         debug!("Start providing {} artifacts", artifact_hashes.len());
         for artifact_hash in artifact_hashes.iter() {
             p2p_client.provide(artifact_hash).await;
@@ -40,12 +41,12 @@ pub async fn provide_artifacts(mut p2p_client: p2p::Client) {
 /// Respond to a RequestArtifact event by getting the artifact from
 /// the ArtifactManager.
 pub async fn handle_request_artifact(
-    mut p2p_client: p2p::Client,
+    mut p2p_client: Client,
     hash: &str,
-    channel: ResponseChannel<p2p::ArtifactResponse>,
+    channel: ResponseChannel<ArtifactResponse>,
 ) {
     let decoded_hash = hex::decode(&hash.get(7..).unwrap()).unwrap();
-    match get_artifact(&decoded_hash, HashAlgorithm::SHA256) {
+    match node_manager::handlers::get_artifact(&decoded_hash, HashAlgorithm::SHA256) {
         Ok(content) => p2p_client.respond_artifact(content, channel).await,
         Err(e) => info!(
             "This node does not provide artifact {}. Error: {:?}",
