+++ src/network/event_loop.rs
@@ -1,0 +1,309 @@
+/*
+   Copyright 2021 JFrog Ltd
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
+use crate::network::artifact_protocol::{ArtifactRequest, ArtifactResponse};
+use crate::network::behaviour::{PyrsiaBehaviour, PyrsiaNetworkEvent};
+use crate::network::client::command::Command;
+use futures::channel::{mpsc, oneshot};
+use futures::prelude::*;
+use libp2p::core::{Multiaddr, PeerId};
+use libp2p::identify::IdentifyEvent;
+use libp2p::kad::{GetClosestPeersOk, GetProvidersOk, KademliaEvent, QueryId, QueryResult};
+use libp2p::multiaddr::Protocol;
+use libp2p::request_response::{
+    RequestId, RequestResponseEvent, RequestResponseMessage, ResponseChannel,
+};
+use libp2p::swarm::SwarmEvent;
+use libp2p::Swarm;
+use log::{debug, info, trace, warn};
+use std::collections::hash_map::Entry::Vacant;
+use std::collections::{HashMap, HashSet};
+use std::error::Error;
+
+type PendingDialMap = HashMap<Multiaddr, oneshot::Sender<Result<(), Box<dyn Error + Send>>>>;
+type PendingListPeersMap = HashMap<QueryId, oneshot::Sender<HashSet<PeerId>>>;
+type PendingStartProvidingMap = HashMap<QueryId, oneshot::Sender<()>>;
+type PendingRequestArtifactMap =
+    HashMap<RequestId, oneshot::Sender<Result<Vec<u8>, Box<dyn Error + Send>>>>;
+
+pub struct PyrsiaEventLoop {
+    swarm: Swarm<PyrsiaBehaviour>,
+    command_receiver: mpsc::Receiver<Command>,
+    event_sender: mpsc::Sender<PyrsiaEvent>,
+    pending_dial: PendingDialMap,
+    pending_list_peers: PendingListPeersMap,
+    pending_start_providing: PendingStartProvidingMap,
+    pending_list_providers: PendingListPeersMap,
+    pending_request_artifact: PendingRequestArtifactMap,
+}
+
+impl PyrsiaEventLoop {
+    pub fn new(
+        swarm: Swarm<PyrsiaBehaviour>,
+        command_receiver: mpsc::Receiver<Command>,
+        event_sender: mpsc::Sender<PyrsiaEvent>,
+    ) -> Self {
+        Self {
+            swarm,
+            command_receiver,
+            event_sender,
+            pending_dial: Default::default(),
+            pending_list_peers: Default::default(),
+            pending_start_providing: Default::default(),
+            pending_list_providers: Default::default(),
+            pending_request_artifact: Default::default(),
+        }
+    }
+
+    pub async fn run(mut self) {
+        loop {
+            futures::select! {
+                event = self.swarm.next() => match event.expect("Swarm stream to be infinite.") {
+                    SwarmEvent::Behaviour(PyrsiaNetworkEvent::Identify(identify_event)) => self.handle_identify_event(identify_event).await,
+                    SwarmEvent::Behaviour(PyrsiaNetworkEvent::Kademlia(kademlia_event)) => self.handle_kademlia_event(kademlia_event).await,
+                    SwarmEvent::Behaviour(PyrsiaNetworkEvent::RequestResponse(request_response_event)) => self.handle_request_response_event(request_response_event).await,
+                    swarm_event => self.handle_swarm_event(swarm_event).await,
+                },
+                command = self.command_receiver.next() => match command {
+                    Some(c) => {
+                        self.handle_command(c).await;
+                    },
+                    // Command channel closed, thus shutting down the network event loop.
+                    None => { warn!("Got empty command"); return },
+                },
+            }
+        }
+    }
+
+    async fn handle_identify_event(&mut self, event: IdentifyEvent) {
+        trace!("Handle IdentifyEvent: {:?}", event);
+        match event {
+            IdentifyEvent::Pushed { .. } => {}
+            IdentifyEvent::Received { peer_id, info } => {
+                println!("Identify::Received: {}; {:?}", peer_id, info);
+                if let Some(addr) = info.listen_addrs.get(0) {
+                    if let Some(sender) = self.pending_dial.remove(addr) {
+                        let _ = sender.send(Ok(()));
+                    }
+
+                    debug!(
+                        "Identify::Received: adding address {:?} for peer {}",
+                        addr.clone(),
+                        peer_id
+                    );
+                    self.swarm
+                        .behaviour_mut()
+                        .kademlia
+                        .add_address(&peer_id, addr.clone());
+                }
+            }
+            IdentifyEvent::Sent { .. } => {}
+            IdentifyEvent::Error { .. } => {}
+        }
+    }
+
+    async fn handle_kademlia_event(&mut self, event: KademliaEvent) {
+        trace!("Handle KademliaEvent: {:?}", event);
+        match event {
+            KademliaEvent::OutboundQueryCompleted {
+                id,
+                result: QueryResult::GetClosestPeers(Ok(GetClosestPeersOk { key: _key, peers })),
+                ..
+            } => {
+                let _ = self
+                    .pending_list_peers
+                    .remove(&id)
+                    .expect("Completed query to be previously pending.")
+                    .send(HashSet::from_iter(peers));
+            }
+            KademliaEvent::OutboundQueryCompleted {
+                id,
+                result: QueryResult::StartProviding(_),
+                ..
+            } => {
+                let sender: oneshot::Sender<()> = self
+                    .pending_start_providing
+                    .remove(&id)
+                    .expect("Completed query to be previously pending.");
+                let _ = sender.send(());
+            }
+            KademliaEvent::OutboundQueryCompleted {
+                id,
+                result:
+                    QueryResult::GetProviders(Ok(GetProvidersOk {
+                        key: _key,
+                        providers,
+                        ..
+                    })),
+                ..
+            } => {
+                let _ = self
+                    .pending_list_providers
+                    .remove(&id)
+                    .expect("Completed query to be previously pending.")
+                    .send(providers);
+            }
+            _ => {}
+        }
+    }
+
+    async fn handle_request_response_event(
+        &mut self,
+        event: RequestResponseEvent<ArtifactRequest, ArtifactResponse>,
+    ) {
+        trace!("Handle RequestResponseEvent: {:?}", event);
+        match event {
+            RequestResponseEvent::Message { message, .. } => match message {
+                RequestResponseMessage::Request {
+                    request, channel, ..
+                } => {
+                    self.event_sender
+                        .send(PyrsiaEvent::RequestArtifact {
+                            hash: request.0,
+                            channel,
+                        })
+                        .await
+                        .expect("Event receiver not to be dropped.");
+                }
+                RequestResponseMessage::Response {
+                    request_id,
+                    response,
+                } => {
+                    let _ = self
+                        .pending_request_artifact
+                        .remove(&request_id)
+                        .expect("Request to still be pending.")
+                        .send(Ok(response.0));
+                }
+            },
+            RequestResponseEvent::InboundFailure { .. } => {}
+            RequestResponseEvent::OutboundFailure {
+                request_id, error, ..
+            } => {
+                let _ = self
+                    .pending_request_artifact
+                    .remove(&request_id)
+                    .expect("Request to still be pending.")
+                    .send(Err(Box::new(error)));
+            }
+            RequestResponseEvent::ResponseSent { .. } => {}
+        }
+    }
+
+    async fn handle_swarm_event(&mut self, event: SwarmEvent<PyrsiaNetworkEvent, impl Error>) {
+        trace!("Handle SwarmEvent: {:?}", event);
+        match event {
+            SwarmEvent::Behaviour(_) => {
+                debug!("Unmatched Behaviour swarm event found: {:?}", event);
+            }
+            SwarmEvent::NewListenAddr { address, .. } => {
+                let local_peer_id = *self.swarm.local_peer_id();
+                info!(
+                    "Local node is listening on {:?}",
+                    address.with(Protocol::P2p(local_peer_id.into()))
+                );
+            }
+            SwarmEvent::ConnectionEstablished { .. } => {}
+            SwarmEvent::ConnectionClosed { .. } => {}
+            SwarmEvent::OutgoingConnectionError { .. } => {}
+            SwarmEvent::BannedPeer { .. } => {}
+            SwarmEvent::Dialing(peer_id) => {
+                debug!(
+                    "Local Peer {} is dialing Peer {}...",
+                    self.swarm.local_peer_id(),
+                    peer_id
+                );
+            }
+            SwarmEvent::ExpiredListenAddr { .. } => {}
+            SwarmEvent::IncomingConnection { .. } => {}
+            SwarmEvent::IncomingConnectionError { .. } => {}
+            SwarmEvent::ListenerClosed { .. } => {}
+            SwarmEvent::ListenerError { .. } => {}
+        }
+    }
+
+    async fn handle_command(&mut self, command: Command) {
+        trace!("Handle Command: {}", command);
+        match command {
+            Command::Listen { addr, sender } => {
+                let _ = match self.swarm.listen_on(addr) {
+                    Ok(_) => sender.send(Ok(())),
+                    Err(e) => sender.send(Err(Box::new(e))),
+                };
+            }
+            Command::Dial { peer_addr, sender } => {
+                if let Vacant(_) = self.pending_dial.entry(peer_addr.clone()) {
+                    match self.swarm.dial(peer_addr.clone()) {
+                        Ok(()) => {
+                            self.pending_dial.insert(peer_addr, sender);
+                        }
+                        Err(e) => {
+                            let _ = sender.send(Err(Box::new(e)));
+                        }
+                    }
+                }
+            }
+            Command::ListPeers { peer_id, sender } => {
+                let query_id = self
+                    .swarm
+                    .behaviour_mut()
+                    .kademlia
+                    .get_closest_peers(peer_id);
+                self.pending_list_peers.insert(query_id, sender);
+            }
+            Command::Provide { hash, sender } => {
+                let query_id = self
+                    .swarm
+                    .behaviour_mut()
+                    .kademlia
+                    .start_providing(hash.into_bytes().into())
+                    .expect("No store error.");
+                self.pending_start_providing.insert(query_id, sender);
+            }
+            Command::ListProviders { hash, sender } => {
+                let query_id = self
+                    .swarm
+                    .behaviour_mut()
+                    .kademlia
+                    .get_providers(hash.into_bytes().into());
+                self.pending_list_providers.insert(query_id, sender);
+            }
+            Command::RequestArtifact { hash, peer, sender } => {
+                let request_id = self
+                    .swarm
+                    .behaviour_mut()
+                    .request_response
+                    .send_request(&peer, ArtifactRequest(hash));
+                self.pending_request_artifact.insert(request_id, sender);
+            }
+            Command::RespondArtifact { artifact, channel } => {
+                self.swarm
+                    .behaviour_mut()
+                    .request_response
+                    .send_response(channel, ArtifactResponse(artifact))
+                    .expect("Connection to peer to be still open.");
+            }
+        }
+    }
+}
+
+#[derive(Debug)]
+pub enum PyrsiaEvent {
+    RequestArtifact {
+        hash: String,
+        channel: ResponseChannel<ArtifactResponse>,
+    },
+}
