+++ src/node_api/handlers/swarm.rs
@@ -15,13 +15,13 @@
 */
 
 use super::{get_config, RegistryError, RegistryErrorCode};
-use crate::network::p2p;
+use crate::network::client::Client;
 use crate::node_manager::{handlers::*, model::cli::Status};
 
 use log::debug;
 use warp::{http::StatusCode, Rejection, Reply};
 
-pub async fn handle_get_peers(mut p2p_client: p2p::Client) -> Result<impl Reply, Rejection> {
+pub async fn handle_get_peers(mut p2p_client: Client) -> Result<impl Reply, Rejection> {
     let peers = p2p_client.list_peers().await;
     debug!("Got received_peers: {:?}", peers);
 
@@ -35,7 +35,7 @@ pub async fn handle_get_peers(mut p2p_client: p2p::Client) -> Result<impl Reply,
         .unwrap())
 }
 
-pub async fn handle_get_status(mut p2p_client: p2p::Client) -> Result<impl Reply, Rejection> {
+pub async fn handle_get_status(mut p2p_client: Client) -> Result<impl Reply, Rejection> {
     let peers = p2p_client.list_peers().await;
 
     let art_count_result = get_arts_count();
